Week 9 - signal Handling

+++ 3_LinuxProgramming/1_process/10_signalContinueStop/main.c
@@ -33,10 +33,13 @@ int main()
         {
             printf("Enter second you want: ");
             scanf("%d", &t);
-            kill(pid, SIGCONT);
-            sleep(t);
-            kill(pid, SIGSTOP);
-        }while(t>0);
+            if(t>0)
+            {
+                kill(pid, SIGCONT);
+                sleep(t);
+                kill(pid, SIGSTOP);
+            }
+        }while(t > 0);
 
 
         kill(pid, SIGKILL);
